Changed string tags to const val
