+++ app/src/main/res/layout/activity_main.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.example.leagueapp.UI.MainActivity">
+    tools:context="com.example.leagueapp.ui.MainActivity">
 
     <FrameLayout
         android:id="@+id/main_framelayout"
